requirements.txt
streamlit
google-generativeai
python-dotenv
openpyxl
pandas
numpy
matplotlib